% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFFTESTING-package.R
\docType{package}
\name{BFFTESTING-package}
\alias{BFFTESTING}
\alias{BFFTESTING-package}
\title{BFFTESTING: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rshudde/BFFTESTING}
  \item Report bugs at \url{https://github.com/rshudde/BFFTESTING/issues}
}

}
\author{
\strong{Maintainer}: rachael shudde \email{rachael.shudde@gmail.com} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
